# chattingapp_be/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      confirm_new_password:
        maxLength: 100
        minLength: 6
        type: string
      new_password:
        maxLength: 100
        minLength: 6
        type: string
      old_password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - confirm_new_password
    - new_password
    - old_password
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        maxLength: 100
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      app_version:
        type: string
      device_name:
        type: string
      device_type:
        type: string
      device_uuid:
        type: string
      identifier:
        type: string
      os_version:
        type: string
      password:
        type: string
      platform:
        type: string
    required:
    - identifier
    - password
    type: object
...
  swaggerdocs.ErrorResponse:
    properties:
      error:
        example: mô tả lỗi
        type: string
      message:
        example: request thất bại
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  swaggerdocs.SuccessOnlyResponse:
    properties:
      message:
        example: thành công
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: |-
    API tài liệu cho ChattingApp Backend.
    Protected APIs dùng Authorization: Bearer <access_token>
  title: ChattingApp Backend API
  version: "1.0"
paths:
  /auth/change-password:
    patch:
      consumes:
      - application/json
      description: Yêu cầu mật khẩu cũ và nhập lại mật khẩu mới 2 lần
      parameters:
      - description: Thông tin đổi mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Đổi mật khẩu
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Gửi email chứa hướng dẫn đặt lại mật khẩu
      parameters:
      - description: Email cần khôi phục
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Quên mật khẩu
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập bằng identifier và password
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.AuthResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Đăng nhập
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Thu hồi refresh token hiện tại
      parameters:
      - description: Refresh token cần thu hồi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/swaggerdocs.LogoutRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Đăng xuất
      tags:
      - Auth
  /auth/me:
    get:
      description: Lấy profile của user đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.UserProfileResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin tài khoản hiện tại
      tags:
      - Auth
  /auth/profile:
    patch:
      consumes:
      - application/json
      description: Cập nhật full_name, avatar_url, bio
      parameters:
      - description: Thông tin profile mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.UserProfileResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật profile của user hiện tại
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Dùng refresh token để lấy access token mới và refresh token mới
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/swaggerdocs.RefreshRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.AuthResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Refresh token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Tạo tài khoản mới
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.AuthResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Đăng ký tài khoản
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Dùng token reset để đặt lại mật khẩu mới
      parameters:
      - description: Token và mật khẩu mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      summary: Đặt lại mật khẩu
      tags:
      - Auth
  /conversations:
    get:
      description: Có hỗ trợ phân trang với page và limit
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số phần tử mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.ConversationListResponseEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách cuộc trò chuyện
      tags:
      - Conversations
  /conversations/{id}:
    get:
      description: Lấy chi tiết conversation theo ID
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.ConversationDetailResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy chi tiết cuộc trò chuyện
      tags:
      - Conversations
  /conversations/{id}/archive:
    patch:
      consumes:
      - application/json
      description: Lưu trữ hoặc bỏ lưu trữ conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Trạng thái archive
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ArchiveConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Archive conversation
      tags:
      - Conversations
  /conversations/{id}/mute:
    patch:
      consumes:
      - application/json
      description: Mute conversation đến thời điểm mute_until hoặc vô thời hạn tùy
        logic service
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin mute
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MuteConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tắt thông báo conversation
      tags:
      - Conversations
  /conversations/{id}/nickname:
    patch:
      consumes:
      - application/json
      description: Cập nhật nickname của user hiện tại trong conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nickname mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateConversationNicknameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật nickname trong conversation
      tags:
      - Conversations
  /conversations/{id}/pin:
    patch:
      consumes:
      - application/json
      description: Ghim hoặc bỏ ghim conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Trạng thái pin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PinConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Pin conversation
      tags:
      - Conversations
  /conversations/{id}/read:
    patch:
      consumes:
      - application/json
      description: Cập nhật last_read_message_id cho thành viên hiện tại
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Dữ liệu đánh dấu đã đọc
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MarkConversationReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Đánh dấu đã đọc conversation
      tags:
      - Conversations
  /conversations/direct:
    post:
      consumes:
      - application/json
      description: Tạo cuộc trò chuyện 1-1 với user khác
      parameters:
      - description: Thông tin tạo direct conversation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDirectConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.ConversationDetailResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo direct conversation
      tags:
      - Conversations
  /conversations/unread-count:
    get:
      description: Lấy tổng số conversation chưa đọc của user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.UnreadCountResponseEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy unread count
      tags:
      - Conversations
  /devices:
    get:
      consumes:
      - application/json
      description: Lấy danh sách thiết bị đã đăng ký của user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.UserDeviceListResponseEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Danh sách thiết bị của tôi
      tags:
      - Devices
  /devices/{uuid}:
    delete:
      consumes:
      - application/json
      description: Disable thiết bị và revoke toàn bộ refresh token của thiết bị đó
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa thiết bị
      tags:
      - Devices
  /devices/{uuid}/disable:
    patch:
      consumes:
      - application/json
      description: Vô hiệu hóa thiết bị và revoke toàn bộ refresh token của thiết
        bị đó
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable thiết bị
      tags:
      - Devices
  /devices/{uuid}/logout:
    post:
      consumes:
      - application/json
      description: Revoke toàn bộ refresh token của thiết bị theo UUID
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout theo thiết bị
      tags:
      - Devices
  /devices/{uuid}/trust:
    patch:
      consumes:
      - application/json
      description: Đánh dấu thiết bị là đáng tin cậy
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trust thiết bị
      tags:
      - Devices
  /devices/{uuid}/untrust:
    patch:
      consumes:
      - application/json
      description: Bỏ đánh dấu thiết bị là đáng tin cậy
      parameters:
      - description: Device UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Untrust thiết bị
      tags:
      - Devices
  /devices/register:
    post:
      consumes:
      - application/json
      description: Đăng ký device cho user hiện tại
      parameters:
      - description: Thông tin device
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Đăng ký hoặc cập nhật device
      tags:
      - Devices
  /e2ee/keys/identity:
    post:
      consumes:
      - application/json
      description: Upload hoặc thay thế active identity public key cho device hiện
        tại. Device phải thuộc user đang đăng nhập.
      parameters:
      - description: Identity key payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UploadIdentityKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.E2EEKeyUploadResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload identity key
      tags:
      - E2EE
  /e2ee/keys/one-time-prekeys:
    post:
      consumes:
      - application/json
      description: Upload danh sách one-time prekeys cho device hiện tại. Các prekey
        này sẽ được consume một lần khi user khác lấy key bundle.
      parameters:
      - description: One-time prekeys payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UploadOneTimePreKeysRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.E2EEKeyUploadResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload one-time prekeys
      tags:
      - E2EE
  /e2ee/keys/signed-prekey:
    post:
      consumes:
      - application/json
      description: Upload hoặc thay thế active signed prekey cho device hiện tại.
        Signed prekey được client ký bằng identity key.
      parameters:
      - description: Signed prekey payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UploadSignedPreKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggerdocs.E2EEKeyUploadResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload signed prekey
      tags:
      - E2EE
  /e2ee/users/{user_id}/key-bundle:
    get:
      description: Lấy identity key, signed prekey và nếu có thì consume một one-time
        prekey của từng active device thuộc user đích.
      parameters:
      - description: Target User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.UserKeyBundleResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy key bundle của user
      tags:
      - E2EE
  /friends:
    get:
      description: Trả về danh sách bạn bè của user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.FriendListResponseEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách bạn bè
      tags:
      - Friends
  /friends/{user_id}:
    delete:
      consumes:
      - application/json
      description: Kết thúc quan hệ bạn bè giữa user hiện tại và user đích
      parameters:
      - description: Target User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggerdocs.SuccessOnlyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hủy kết bạn
      tags:
      - Friends
...
  /friends/requests/incoming:
    get:
      consumes:
      - application/json
      description: Lấy danh sách lời mời kết bạn đang chờ mà người hiện tại là người
        nhận
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FriendRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggerdocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Danh sách lời mời đến
      tags:
      - Friends
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Nhập theo dạng: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"